% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GExDesView.R
\name{GExpDesView}
\alias{GExpDesView}
\title{GExpDes: interface Gráfica para ExpDes}
\usage{
GExpDesView()
}
\value{
Abre uma aplicação Shiny
}
\description{
Abre um ambiente onde se inicia o processo de análise estatística
}
\examples{
require(GExpDes)
GExpDesView()
}
\author{
Rogerio Kormann, Eduardo Nunes Rosa, Crysttian Arantes Paixao,Eric Batista Ferreira e Denismar Alves Nogueira
}
